refactor: introduce core and utils packages
