+++ tests/test_planning.py
@@ -1,5 +1,5 @@
 from assistant import WorkAssistant, LLMClient
-from session_manager import SessionManager
+from core.session_manager import SessionManager
 from unittest.mock import MagicMock, call
