+++ tests/test_work_patterns.py
@@ -1,7 +1,7 @@
 import openai
 from datetime import datetime
 from assistant import LLMClient, Ticket
-from session_manager import SessionManager
+from core.session_manager import SessionManager
 
 
 def test_work_patterns_persist(tmp_path):
